New directory structure. Version bump to v0.0.44 (#141)
Pypechain generated output types now wraps all files in an outer
directory. This is necessary for out of memory issues with `pyright` on
large solidity packages.

Breaking changes:
- Imports now need to point to the contract name.
`from types import SolidityFileContract` -> `from types.SolidityFile
import SolidityFileContract`

New changes:
- Imports of types are now exposed in the directory.
`from types.SolidityFileTypes import MyType` -> `from types.SolidtyFile
import MyType` or `from types import SolidityFile; SolidityFile.MyType`
- Adding guard to clearing output directory by refusing to remove all in
output directory if not empty and not determined to be pypechain output
dir. Solves https://github.com/delvtech/pypechain/issues/140
- Explicit gas estimate from pending block in `sign_and_transact`.

+++ pypechain/main.py
@@ -11,9 +11,7 @@ from pathlib import Path
 from typing import NamedTuple, Sequence
 
 from pypechain.render import render_contracts
-from pypechain.render.init import render_init_file
 from pypechain.utilities.abi import AbiInfo, load_abi_infos_from_file
-from pypechain.utilities.types import RenderOutput
 
 
 def main(argv: Sequence[str] | None = None) -> None:
@@ -72,17 +70,18 @@ def pypechain(
     # Create/clear the output directory
     setup_directory(output_dir)
 
-    # Now process all gathered files
-    file_outputs: list[RenderOutput] = render_contracts(
-        abi_infos, output_dir, line_length, apply_formatting, parallel=parallel
-    )
-
-    # Render the __init__.py file
-    render_init_file(output_dir, file_outputs, line_length)
-
+    # Since setup directory looks for `pypechain.version`, we make this file first
     # Make a pypechain.version file
     with open(f"{output_dir}/pypechain.version", "w", encoding="utf-8") as f:
         f.write(f"pypechain == {importlib.metadata.version('pypechain')}")
+
+    # Now process all gathered files
+    render_contracts(abi_infos, output_dir, line_length, apply_formatting, parallel=parallel)
+
+    # Make an empty __init__.py file
+    # render_init_file(output_dir, file_outputs, line_length)
+    with open(f"{output_dir}/__init__.py", "w", encoding="utf-8") as f:
+        pass
 
 
 def gather_json_files(directory: str) -> list:
@@ -93,12 +92,20 @@ def gather_json_files(directory: str) -> list:
 def setup_directory(directory: str) -> None:
     """Set up the output directory. If it exists, clear it. Otherwise, create it."""
 
-    # If the directory exists, remove it
+    print(f"Setting up output directory {directory}.")
+
+    # If the directory exists, make sure it's a pypechain output directory
     if os.path.exists(directory):
-        shutil.rmtree(directory)
+        if os.path.isfile(f"{directory}/pypechain.version"):
+            # Clear the directory if it's a pypechain output directory
+            shutil.rmtree(directory)
+        else:
+            # Print error if directory isn't empty
+            if not len(os.listdir(directory)) == 0:
+                raise ValueError(f"Directory {directory} isn't empty, and is not a pypechain output directory.")
 
     # Create the directory
-    os.makedirs(directory)
+    os.makedirs(directory, exist_ok=True)
 
 
 class Args(NamedTuple):
